backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pytesseract
pdf2image
camelot-py[cv]
opencv-python
pillow
PyMuPDF
pdfplumber
python-multipart
xgboost
scikit-learn
numpy
pandas
joblib
fastapi[all]
nltk
requests
huggingface_hub>=0.26.0